# internal/api/spec/knot-openapi.yaml
openapi: 3.0.3
servers:
  - url: https://127.0.0.1
    description: Local server.
info:
  version: 1.0.0
  title: Knot API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: ../images/logo-light.svg
  description: |
    # Introduction
    The REST API is used for the CLI, Web interface and 3rd party integrations.
...
paths:
...
  /api/tunnels/{tunnel_name}:
    delete:
      tags:
        - Tunnels
      summary: Delete Tunnel
      description: Delete a specific tunnel for the authenticated user.
      operationId: deleteTunnel
      parameters:
        - name: tunnel_name
          in: path
          required: true
          description: The name of the tunnel to delete.
          schema:
            type: string
      responses:
        '200':
          description: Tunnel deleted successfully.
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/not-found'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message.
...
  responses:
...
    not-found:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    internal-server-error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...